# apps/backend/docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/iam/assignments": {
      "post": {
        "operationId": "RoleAssignmentController_assign_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleAssignmentDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Assign a role (with scope) to a user — idempotent for an identical active assignment",
        "tags": [
          "IAM"
        ]
      }
    },
    "/api/v1/iam/assignments/{id}": {
      "delete": {
        "operationId": "RoleAssignmentController_revoke_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Revoke a role assignment",
        "tags": [
          "IAM"
        ]
      }
    },
    "/api/v1/iam/users/{userId}/assignments": {
      "get": {
        "operationId": "RoleAssignmentController_listForUser_v1",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RoleAssignmentDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List a user's role assignments (active and revoked)",
        "tags": [
          "IAM"
        ]
      }
    },
...
    "/api/v1/provider-onboarding/applications/me": {
      "post": {
        "operationId": "ApplicationController_submit_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApplicationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationDto"
                }
              }
            }
          },
          "404": {
            "description": "No provider profile yet, or referredByAgentCode is not recognized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "A non-rejected application already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Submit an onboarding application for the current provider, or resubmit a previously REJECTED one",
        "tags": [
          "Provider Onboarding"
        ]
      },
      "get": {
        "operationId": "ApplicationController_get_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get the current provider's application, with its documents",
        "tags": [
          "Provider Onboarding"
        ]
      }
    },
    "/api/v1/provider-onboarding/applications/me/documents": {
      "post": {
        "operationId": "ApplicationController_addDocument_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDocumentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Attach a document to the current provider's application",
        "tags": [
          "Provider Onboarding"
        ]
      },
      "get": {
        "operationId": "ApplicationController_listDocuments_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DocumentDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List the current provider's application documents",
        "tags": [
          "Provider Onboarding"
        ]
      }
    },
...
    "/api/v1/payments/me": {
      "post": {
        "operationId": "CustomerPaymentController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePaymentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentDto"
                }
              }
            }
          },
          "404": {
            "description": "Booking not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Booking cannot accept payment, or the amount exceeds the outstanding balance",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Initiate a payment against one of your own bookings",
        "tags": [
          "Payment"
        ]
      },
      "get": {
        "operationId": "CustomerPaymentController_list_v1",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "INITIATED",
                "SUCCEEDED",
                "FAILED"
              ]
            }
          },
          {
            "name": "bookingId",
            "required": false,
            "in": "query",
            "description": "Restrict to payments against a specific booking",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PaymentDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List payments you have made",
        "tags": [
          "Payment"
        ]
      }
    },
    "/api/v1/payments/me/{id}": {
      "get": {
        "operationId": "CustomerPaymentController_findOne_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get one of your own payments",
        "tags": [
          "Payment"
        ]
      }
    },
    "/api/v1/payments/me/{id}/verify": {
      "post": {
        "operationId": "CustomerPaymentController_verify_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyPaymentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Payment is not ONLINE, or is not in INITIATED status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Confirm an ONLINE payment after completing checkout with the gateway",
        "tags": [
          "Payment"
        ]
      }
    },
...
  },
  "info": {
    "title": "SevaMitra API",
    "description": "SevaMitra backend API — see docs/ARCHITECTURE.md",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ErrorDetailDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "BAD_REQUEST"
          },
          "message": {
            "type": "string",
            "example": "Incorrect OTP"
          },
          "details": {
            "type": "object"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetailDto"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "AssignRoleDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
          },
          "roleId": {
            "type": "string",
            "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
          },
          "scopeType": {
            "type": "string",
            "enum": [
              "PLATFORM",
              "GEOGRAPHY",
              "ORGANIZATION",
              "PROVIDER_GROUP",
              "PROVIDER_COMPANY",
              "OPERATIONAL_AREA"
            ],
            "default": "PLATFORM"
          },
          "scopeId": {
            "type": "string",
            "example": null,
            "nullable": true,
            "description": "Opaque reference into another module's entity (e.g. a taluk id). Required unless scopeType=PLATFORM."
          }
        },
        "required": [
          "userId",
          "roleId",
          "scopeType"
        ]
      },
      "RoleAssignmentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
          },
          "userId": {
            "type": "string",
            "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
          },
          "roleId": {
            "type": "string",
            "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
          },
          "roleName": {
            "type": "string",
            "example": "Taluk Ops Manager"
          },
          "scopeType": {
            "type": "string",
            "enum": [
              "PLATFORM",
              "GEOGRAPHY",
              "ORGANIZATION",
              "PROVIDER_GROUP",
              "PROVIDER_COMPANY",
              "OPERATIONAL_AREA"
            ],
            "example": "PLATFORM"
          },
          "scopeId": {
            "type": "object",
            "example": null,
            "nullable": true
          },
          "assignedBy": {
            "type": "object",
            "example": null,
            "nullable": true
          },
          "assignedAt": {
            "format": "date-time",
            "type": "string"
          },
          "revokedAt": {
            "type": "object",
            "example": null,
            "nullable": true
          }
        },
        "required": [
          "id",
          "userId",
          "roleId",
          "roleName",
          "scopeType",
          "assignedAt"
        ]
      },
...
      "CreateApplicationDto": {
        "type": "object",
        "properties": {
          "referredByAgentCode": {
            "type": "string",
            "example": "A1B2C3D4",
            "description": "An existing agent's agentCode — sets channel to AGENT_REFERRED and attributes this application to them"
          }
        }
      },
      "DocumentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
          },
          "applicationId": {
            "type": "string",
            "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
          },
          "type": {
            "type": "string",
            "enum": [
              "IDENTITY",
              "ADDRESS_PROOF",
              "BUSINESS_REGISTRATION",
              "SKILL_CERTIFICATE",
              "LICENSE",
              "REFERENCE",
              "OTHER"
            ],
            "example": "IDENTITY"
          },
          "fileUrl": {
            "type": "string",
            "example": "https://example.com/uploads/aadhaar-front.jpg"
          },
          "label": {
            "type": "object",
            "example": "Aadhaar card, front",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "applicationId",
          "type",
          "fileUrl",
          "createdAt"
        ]
      },
      "ApplicationDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
          },
          "providerId": {
            "type": "string",
            "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
          },
          "channel": {
            "type": "string",
            "enum": [
              "SELF",
              "AGENT_REFERRED"
            ],
            "example": "SELF"
          },
          "referredByAgentId": {
            "type": "object",
            "example": null,
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "SUBMITTED",
              "UNDER_REVIEW",
              "APPROVED",
              "REJECTED"
            ],
            "example": "SUBMITTED"
          },
          "reviewNote": {
            "type": "object",
            "example": null,
            "nullable": true
          },
          "reviewedBy": {
            "type": "object",
            "example": null,
            "nullable": true
          },
          "submittedAt": {
            "format": "date-time",
            "type": "string"
          },
          "reviewedAt": {
            "type": "object",
            "example": null,
            "nullable": true
          },
          "documents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DocumentDto"
            }
          }
        },
        "required": [
          "id",
          "providerId",
          "channel",
          "status",
          "submittedAt"
        ]
      },
      "CreateDocumentDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "IDENTITY",
              "ADDRESS_PROOF",
              "BUSINESS_REGISTRATION",
              "SKILL_CERTIFICATE",
              "LICENSE",
              "REFERENCE",
              "OTHER"
            ],
            "example": "IDENTITY"
          },
          "fileUrl": {
            "type": "string",
            "example": "https://example.com/uploads/aadhaar-front.jpg",
            "description": "A reference URL — no upload endpoint exists yet (Media/Document module is Phase 1b), this is trusted as given for now"
          },
          "label": {
            "type": "string",
            "example": "Aadhaar card, front"
          }
        },
        "required": [
          "type",
          "fileUrl"
        ]
      },
...
      "CreatePaymentDto": {
        "type": "object",
        "properties": {
          "bookingId": {
            "type": "string",
            "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
          },
          "method": {
            "type": "string",
            "enum": [
              "ONLINE",
              "CASH"
            ],
            "example": "ONLINE"
          },
          "amount": {
            "type": "number",
            "example": 499,
            "description": "Amount being paid in this transaction — may be partial (an advance)"
          }
        },
        "required": [
          "bookingId",
          "method",
          "amount"
        ]
      },
      "PaymentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
          },
          "bookingId": {
            "type": "string",
            "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
          },
          "method": {
            "type": "string",
            "enum": [
              "ONLINE",
              "CASH"
            ],
            "example": "ONLINE"
          },
          "amount": {
            "type": "string",
            "example": "499.00"
          },
          "currency": {
            "type": "string",
            "example": "INR"
          },
          "status": {
            "type": "string",
            "enum": [
              "INITIATED",
              "SUCCEEDED",
              "FAILED"
            ],
            "example": "INITIATED"
          },
          "gatewayOrderId": {
            "type": "object",
            "nullable": true
          },
          "gatewayPaymentId": {
            "type": "object",
            "nullable": true
          },
          "failureReason": {
            "type": "object",
            "nullable": true
          },
          "settledAt": {
            "type": "object",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "bookingId",
          "method",
          "amount",
          "currency",
          "status",
          "createdAt",
          "updatedAt"
        ]
      },
      "VerifyPaymentDto": {
        "type": "object",
        "properties": {
          "gatewayPaymentId": {
            "type": "string",
            "example": "stub_pay_1234567890"
          },
          "gatewaySignature": {
            "type": "string",
            "example": "a1b2c3d4e5f6..."
          }
        },
        "required": [
          "gatewayPaymentId",
          "gatewaySignature"
        ]
      },
...
    }
  }
}